Add necessary skeletal code files
The essential manifest.json file has been added and initialised with
the appropriate fields.

The content-script.js file has been added and initialised with an Event
Listener for text selection that logs selected text to the console.

The HTML and CSS files for the extension's popup has also been added.

In its current state, Pithy is able to be added to and run on Chrome.